reorder result object and split files
